# swagger-docs/swagger.yaml
openapi: 3.0.0
info:
  title: learn express and swagger
  description: a course segment about writting docs
  version: 1.0.0
  contact:
    email: Radhey@gmail.com
    # phone: 1234567890
    url: https://lco.dev

# servers:
#   - url: "https://localhost:3000/api/v1"
#     description: for localhost - secure
#   - url: "http://localhost:3000/api/v1"
#     description: for localhost - regular

servers:
  - url: "{protocol}://localhost:3000/api/{version}"
    description: for localhost - secure
    variables:
      version:
        enum:
          - v1
          - v2
      protocol:
        enum:
          - http
          - https

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer

paths:
  /lco:
    get:
      tags:
        - String
      summary: return a greet message from lco
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: string
                example: "mystring"
        400:
          description: Bad Request
        500:
          description: Internal Server Error

  /lcoobject:
    get:
      tags:
        - Object
      summary: return lco courses
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        400:
          description: Bad Request
        500:
          description: Internal Server Error

  /courses:
    get:
      tags:
        - Array
      summary: return all courses
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: number
        400:
          description: Bad Request
        500:
          description: Internal Server Error

  /mycourse/{courseId}:
    get:
      tags:
        - String
      summary: return course based on request Id
      parameters:
        - name: courseId
          in: path
          required: true
          default: 22
          schema:
            type: string
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        400:
          description: Bad Request
        500:
          description: Internal Server Error

  /addCourse:
    post:
      tags:
        - String
      summary: add course in array
      consumes:
        - application/json
      produces:
        - application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                price:
                  type: number

      # parameters:
      #   - name: courseId
      #     in: path
      #     required: true
      #     default: 22
      #     schema:
      #       type: string
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        400:
          description: Bad Request
        500:
          description: Internal Server Error
  /courseQuery:
    get:
      tags:
        - String
      summary: trying to learn about query
      parameters:
        - name: location
          in: query
          required: true
          schema:
            type: string
            enum: [delhi, mumbai, jaipur]
        - name: device
          in: query
          required: true
          schema:
            type: string
            enum: [web, mobile]
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    type: string
                  device:
                    type: string
        400:
          description: Bad Request
        500:
          description: Internal Server Error

  /courseUpload:
    post:
      tags:
        - String
      summary: upload course images
      parameters:
        - in: header
          name: auth
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: All good success
        400:
          description: Bad Request
        500:
          description: Internal Server Error